Add the ability to resolve locally
